Usuario x Perfil

+++ pep-web/src/main/java/br/edu/utfpr/service/PerfilService.java
@@ -8,6 +8,7 @@ import javax.inject.Named;
 import javax.persistence.NoResultException;
 
 import br.edu.utfpr.dao.PerfilDao;
+import br.edu.utfpr.exception.AppException;
 import br.edu.utfpr.model.Perfil;
 
 @Named
@@ -21,11 +22,24 @@ public class PerfilService {
 		return perfilDao.retornarPerfis();
 	}
 	
-	public void removerPerfil(Perfil perfil) {
+	public void removerPerfil(Perfil perfil) throws AppException {
+		perfil = perfilDao.getById(perfil.getId());
+		
+		//se possui usuarios vinculados, nao permite excluir perfil
+		if (!perfil.getUsuarios().isEmpty()) {
+			throw new AppException("perfil.remover.erro.usuariosvinculados");
+		}
+		
 		perfilDao.removeById(perfil.getId());
 	}
 	
-	public void salvarPerfil(Perfil perfil) {
+	public void salvarPerfil(Perfil perfil) throws AppException {
+		//valida se nao esta duplicando
+		Perfil perfilBase = this.retornarPerfilPorNome(perfil.getNome());
+		if (perfilBase != null && !perfilBase.getId().equals(perfil.getId())) {
+			throw new AppException("perfil.salvar.erro.nomeexistente", perfil.getNome());
+		}
+		
 		perfilDao.save(perfil);
 	}
